Disable "maybe-uninitialized" warning globally
Several build configurations had already disabled this warning because
it generates a lot of false positives. But some had not, and it was
still enabled for "allmodconfig" builds, for example.

Looking at the warnings produced, every single one I looked at was a
false positive, and the warnings are frequent enough (and big enough)
that they can easily hide real problems that you don't notice in the
noise generated by -Wmaybe-uninitialized.

The warning is good in theory, but this is a classic case of a warning
that causes more problems than the warning can solve.

If gcc gets better at avoiding false positives, we may be able to
re-enable this warning. But as is, we're better off without it, and I
want to be able to see the *real* warnings.

Change-Id: Ie810d255be8911c413c9abe6965a9a66639a1dce
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

* Adjust qcacld-3.0 accordingly

../drivers/staging/qcacld-3.0/core/wma/src/wma_scan_roam.c: In function 'wma_roam_scan_fill_self_caps':
../drivers/staging/qcacld-3.0/core/wma/src/wma_scan_roam.c:3521:26: warning: 'selfCaps.shortSlotTime' may be used uninitialized in this function [-Wmaybe-uninitialized]
 3521 |   selfCaps.shortSlotTime = 1;
      |   ~~~~~~~~~~~~~~~~~~~~~~~^~~
../drivers/staging/qcacld-3.0/core/wma/src/wma_scan_roam.c:3528:24: warning: 'selfCaps.spectrumMgt' may be used uninitialized in this function [-Wmaybe-uninitialized]
 3528 |   selfCaps.spectrumMgt = 1;
      |   ~~~~~~~~~~~~~~~~~~~~~^~~
../drivers/staging/qcacld-3.0/core/wma/src/wma_scan_roam.c:3535:16: warning: 'selfCaps.qos' may be used uninitialized in this function [-Wmaybe-uninitialized]
 3535 |   selfCaps.qos = 1;
      |   ~~~~~~~~~~~~~^~~
../drivers/staging/qcacld-3.0/core/wma/src/wma_scan_roam.c:3543:17: warning: 'selfCaps.apsd' may be used uninitialized in this function [-Wmaybe-uninitialized]
 3543 |   selfCaps.apsd = 1;
      |   ~~~~~~~~~~~~~~^~~

Signed-off-by: Ahmad Thoriq Najahi <najahiii@outlook.co.id>
Signed-off-by: Sujit Roy <kumarsujitroy@gmail.com>